# lib/api-zod/src/generated/types/createNoticeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateNoticeBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  body: string;
  /** @minLength 1 */
  category: string;
  targetRole?: string;
  postedAt?: Date;
}