# lib/api-zod/src/generated/types/notice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Notice {
  id: string;
  title: string;
  body: string;
  category: string;
  targetRole: string;
  postedAt: Date;
  isRead: boolean;
  createdAt: Date;
}